requirements.txt
Flask==3.2.1
psycopg[binary]==3.1.7
Werkzeug==2.3.7
python-dotenv==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
click==8.1.7
gunicorn==21.2.0
